fixed an issue where editing a product would allow for 2 products with the same id

+++ InventorySystemMain.java
@@ -79,10 +79,14 @@ public class InventorySystemMain {
             Product original = Database.getProductById(id, db.products).get(0);
             ProductUI dialog = new ProductUI("Edit Product", original);
             if (dialog.product != null) {   // check if input was validated and a valid product was returned
-                db.removeProduct(original);
-                db.addProduct(dialog.product);
-                ui.updateRows(db);
-                ui.log("Edit succeeded.");
+                if (Database.getProductById(dialog.product.getId(), db.products).size() > 0 || dialog.product.getId() < 1)
+                    ui.log("Id already in use or Id negative, product edit failed.");
+                else {
+                    db.removeProduct(original);
+                    db.addProduct(dialog.product);
+                    ui.updateRows(db);
+                    ui.log("Edit succeeded.");
+                }
             }
             else
                 ui.log("Edit canceled by user.");
